refactor(Tooltip): BalloonをControlledTooltipにリネームし、概念を整理
- Balloon.tsx → ControlledTooltip.tsx にリネーム
- Balloon/Tooltip/ → Tooltip/ にディレクトリを移動
- Balloonディレクトリを削除
- src/index.ts で ControlledTooltip as Balloon として後方互換性を維持
- @deprecated はBalloonのみに適用（Tooltipは非推奨ではない）
- すべての内部コンポーネントのimportパスを更新
- ThCheckboxでBalloon → ControlledTooltipに変更

概念の整理:
- ControlledTooltip: 吹き出しの基本実装（旧Balloon）
- Tooltip: ControlledTooltipを使った使いやすいツールチップ
- Balloon: ControlledTooltipのalias（後方互換・非推奨・Tour用途のみ）
