--- packages/smarthr-ui/src/components/Balloon/Tooltip/index.ts
@@ -1,1 +1,0 @@
-export { Tooltip } from './Tooltip'
