--- packages/smarthr-ui/src/components/Balloon/index.ts
@@ -1,2 +1,0 @@
-export { Balloon } from './Balloon'
-export { Tooltip } from './Tooltip'
