+++ packages/smarthr-ui/src/components/LineClamp/LineClamp.tsx
@@ -11,7 +11,7 @@ import {
 } from 'react'
 import { type VariantProps, tv } from 'tailwind-variants'
 
-import { Tooltip } from '../Balloon'
+import { Tooltip } from '../Tooltip'
 
 type AbstractProps = PropsWithChildren<VariantProps<typeof classNameGenerator>>
 type Props = AbstractProps & Omit<ComponentPropsWithRef<'span'>, keyof AbstractProps>
