+++ packages/smarthr-ui/src/components/Tooltip/index.ts
@@ -1,0 +1,2 @@
+export { ControlledTooltip } from './ControlledTooltip'
+export { Tooltip } from './Tooltip'
