+++ packages/smarthr-ui/src/components/Tooltip/tooltipHelper.ts
