+++ packages/smarthr-ui/src/index.ts
@@ -2,12 +2,6 @@ import './configureTwMerge'
 
 // components
 export { DisclosureTrigger, DisclosureContent } from './components/Disclosure'
-/**
- * @deprecated 通常の用途では Tooltip コンポーネントを使用してください。
- * Tour（アプリの初回利用時チュートリアル）のような特殊な用途でのみ使用可能ですが、
- * 将来的には Tour 専用のコンポーネントとして整理される予定です。
- */
-export { Balloon } from './components/Balloon'
 export { Checkbox } from './components/Checkbox'
 export { Chip } from './components/Chip'
 export {
@@ -86,7 +80,13 @@ export {
   AccordionPanelTrigger,
 } from './components/AccordionPanel'
 export { InformationPanel } from './components/InformationPanel'
-export { Tooltip } from './components/Balloon'
+/**
+ * @deprecated 通常の用途では Tooltip コンポーネントを使用してください。
+ * Tour（アプリの初回利用時チュートリアル）のような特殊な用途でのみ使用可能ですが、
+ * 将来的には Tour 専用のコンポーネントとして整理される予定です。
+ */
+export { ControlledTooltip as Balloon } from './components/Tooltip'
+export { Tooltip } from './components/Tooltip'
 export { BottomFixedArea } from './components/BottomFixedArea'
 export {
   ErrorScreen,
